test(implementation): extracted all tests modules
- extracted all `tests` module to individual files.
- edited `NekosLifeError` to use in unittest
