+++ src/implementation.rs
@@ -127,73 +127,7 @@ pub async fn get(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use strum::IntoEnumIterator;
-
-    #[tokio::test]
-    async fn all_endpoints_work() {
-        let client = reqwest::Client::new();
-
-        for variant in Category::iter() {
-            get_with_client(&client, variant)
-                .await
-                .unwrap_or_else(|_| {
-                    panic!(
-                        "{} does not work",
-                        variant.to_url_path()
-                    )
-                });
-            println!("{}: works", variant.to_url_path());
-        }
-    }
-
-    #[tokio::test]
-    async fn no_new_endpoints() {
-        use {
-            regex::Regex, std::collections::HashSet,
-            strum::IntoEnumIterator,
-        };
-
-        let regex_img =
-            Regex::new(r"^GET,HEAD,OPTIONS\s+/api/v2/img/<(?P<eps>.*)>$")
-                .expect("failed to init regex");
-
-        assert_eq!(
-            Category::iter()
-                .map(|c| c.to_url_path())
-                .chain(["v3", "nekoapi_v3.1"])
-                .collect::<HashSet<_>>(),
-            Regex::new(r"'(?P<ct>[\w\.]+)'")
-                .expect("failed to init regex")
-                .captures_iter(
-                    reqwest::Client::new()
-                        .get(
-                            BASEURL.join("endpoints").expect(
-                                "an error occurred while joining the url"
-                            )
-                        )
-                        .send()
-                        .await
-                        .expect("failed to send request")
-                        .json::<Vec<String>>()
-                        .await
-                        .expect("failed to parse response")
-                        .iter()
-                        .find_map(|line| regex_img.captures(line))
-                        .expect("no match found")
-                        .name("eps")
-                        .expect("couldn't find capture named eps")
-                        .as_str()
-                )
-                .map(|cap| cap
-                    .name("ct")
-                    .expect("couldn't find capture named ct")
-                    .as_str())
-                .collect::<HashSet<_>>()
-        );
-    }
-}
+mod tests;
 
 /// blocking version of implementation
 #[cfg(feature = "blocking")]
